+++ src/node-lib/curriculum-api-2023/__mocks__/index.ts
@@ -1,5 +1,6 @@
 import unitListingFixture from "../fixtures/unitListing.fixture";
 import { unitBrowseDataFixture } from "../fixtures/unitBrowseData.fixture";
+import { subjectBrowseDataFixture } from "../fixtures/subjectBrowseData.fixture";
 
 import { specialistSubjectListingFixture2023 } from "@/node-lib/curriculum-api-2023/fixtures/specialistSubjectListing.fixture";
 import programmeListingFixture from "@/node-lib/curriculum-api-2023/fixtures/programmeListing.fixture";
@@ -16,6 +17,7 @@ import { lessonBrowseDataFixture } from "@/node-lib/curriculum-api-2023/fixtures
 import lessonShareFixtures from "@/node-lib/curriculum-api-2023/fixtures/lessonShare.fixture";
 import lessonDownloadsFixtures from "@/node-lib/curriculum-api-2023/fixtures/lessonDownloads.fixture";
 import { subjectListingFixture2023 } from "@/node-lib/curriculum-api-2023/fixtures/subjectListing.fixture";
+import { pupilProgrammeListingFixture } from "@/node-lib/curriculum-api-2023/fixtures/pupilProgrammeListing.fixture";
 import { mockUrls } from "@/node-lib/curriculum-api-2023/fixtures/teachersSiteMap.fixture";
 
 const curriculumApi: Pick<
@@ -31,6 +33,8 @@ const curriculumApi: Pick<
   | "pupilLessonQuery"
   | "pupilLessonListingQuery"
   | "pupilUnitListingQuery"
+  | "pupilSubjectListingQuery"
+  | "pupilProgrammeListingQuery"
   | "specialistSubjectListing"
   | "specialistUnitListing"
   | "specialistLessonOverviewCanonical"
@@ -69,6 +73,12 @@ const curriculumApi: Pick<
   }),
   pupilUnitListingQuery: jest.fn(async () => {
     return [unitBrowseDataFixture({})];
+  }),
+  pupilSubjectListingQuery: jest.fn(async () => {
+    return [subjectBrowseDataFixture({})];
+  }),
+  pupilProgrammeListingQuery: jest.fn(async () => {
+    return pupilProgrammeListingFixture();
   }),
   lessonShare: jest.fn(async () => {
     return lessonShareFixtures();