+++ src/pages/pupils/beta/programmes/[programmeSlug]/options-l.tsx
@@ -1,0 +1,39 @@
+// TESTING
+// http://localhost:3000/pupils/beta/programmes/maths-secondary-year-10/options-l
+// CF http://localhost:3000/pupils/beta/programmes/maths-secondary-year-10/options
+
+import { GetStaticProps } from "next";
+
+import { PupilViewsProgrammeListing } from "@/components/PupilViews/PupilProgrammeListing/PupilProgrammeListing.view";
+import {
+  ProgrammesPageProps,
+  OptionsURLParams,
+  getPupilOptionData,
+} from "@/pages-helpers/pupil/options-pages/options-pages-helpers";
+import { getStaticPaths as getStaticPathsTemplate } from "@/pages-helpers/get-static-paths";
+
+const ProgrammesPage = ({
+  programmes,
+  baseSlug,
+  yearSlug,
+}: ProgrammesPageProps) => {
+  return (
+    <PupilViewsProgrammeListing
+      programmes={programmes}
+      baseSlug={baseSlug}
+      isLegacy={true}
+      yearSlug={yearSlug}
+    />
+  );
+};
+
+export const getStaticPaths = getStaticPathsTemplate<OptionsURLParams>;
+
+export const getStaticProps: GetStaticProps<
+  ProgrammesPageProps,
+  OptionsURLParams
+> = async (context) => {
+  return getPupilOptionData(context, true);
+};
+
+export default ProgrammesPage;